+++ packages/expo-ui/android/src/main/java/expo/modules/ui/ExpoUIModule.kt
@@ -44,7 +44,7 @@ class ExpoUIModule : Module() {
     View(ProgressView::class)
 
     View(TextInputView::class)  {
-      Events("onValueChanged")
+      Events("onValueChanged", "onTextFieldFocus", "onTextFieldBlur")
     }
   }
 }
